requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
pytz
sqlparse
tzdata